Add Double Jump Events (#17)
* 1.16.3 and CraftBukkit Script Fix

* CraftBukkit Script Fix

* CraftBukkit Script Fix Line Seperators

* Remove MVdWPAPI

The website it was hosted on was down and already had pre-existing problems, so I just removed it.

* Added events

+ DoubleJumpEvent (for manipulating variables for the double jump)
+ GroundPoundEvent (for manipulating the variables for the ground pound)
+ PreDoubleJumpEvent (can cancel double jumping with this event)

* Fixed some other files I realized needed updating

* Removed unneeded class and attempted to cache files

* Fix cache

* Update codeql-analysis.yml

+++ NmsReference/src/main/java/me/treyruffy/treysdoublejump/nmsreference/ParticlesMain.java
@@ -9,7 +9,7 @@ import org.bukkit.entity.Player;
 
 public interface ParticlesMain {
 
-	// Interface for all versions 1.8 - 1.15
-	void sendParticle(Player p, String particle, Location loc, int amount, int r, int g, int b);
+	// Interface for all versions 1.8 - 1.16.4+
+	void sendParticle(Player p, String particle, Location loc, int amount, float r, float g, float b);
 	
 }
